+++ src/kate.shared.CommandLine/CommandActionAttribute.cs
@@ -1,5 +1,5 @@
 ﻿/*
-   Copyright 2022-2025 Kate Ward <kate@dariox.club>
+   Copyright 2022-2026 Kate Ward <kate@dariox.club>
 
    Licensed under the Apache License, Version 2.0 (the "License");
    you may not use this file except in compliance with the License.
@@ -15,12 +15,14 @@
 */
 
 using System;
+using JetBrains.Annotations;
 
 namespace kate.shared.CommandLine
 {
     /// <summary>
     /// Required attribute about details for <see cref="IAction"/>
     /// </summary>
+    [PublicAPI]
     [AttributeUsage(AttributeTargets.Class, AllowMultiple = false)]
     public class CommandActionAttribute : Attribute
     {
@@ -28,6 +30,7 @@ namespace kate.shared.CommandLine
         /// Name of the action to use when invoking via command-line.
         /// </summary>
         public string ActionName { get; set; }
+        
         /// <summary>
         /// Type of the class where all the action-specific options are.
         /// </summary>